# public/openapi.yaml
openapi: 3.0.1
info:
  version: 1.0.0
  title: Businesses API
  description: A simple API for a Yelp-like application.
paths:
...
  /users/{userID}/businesses:
    parameters:
      - name: userID
        in: path
        description: Unique ID of a user.
        schema:
          type: integer
        example: 0
        required: true
    get:
      summary: Fetch a user's owned businesses.
      description: >
        Returns a list of all of the businesses owned by the user specified by `userID`.
      operationId: getUserBusinesses
      tags:
        - users
      responses:
        200:
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  lodgings:
                    type: array
                    description: The businesses owned by the specified user.
                    items:
                      $ref: '#/components/schemas/Business'
  /users/{userID}/reviews:
    parameters:
      - name: userID
        in: path
        description: Unique ID of a user.
        schema:
          type: integer
        example: 21
        required: true
    get:
      summary: Fetch a user's reviews.
      description: >
        Returns a list of all of the reviews written by the user specified by `userID`.
      operationId: getUserReviews
      tags:
        - users
      responses:
        200:
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  lodgings:
                    type: array
                    description: The reviews written by the specified user.
                    items:
                      $ref: '#/components/schemas/Review'
  /users/{userID}/photos:
    parameters:
      - name: userID
        in: path
        description: Unique ID of a user.
        schema:
          type: integer
        example: 21
        required: true
    get:
      summary: Fetch a user's photos.
      description: >
        Returns a list of all of the photos posted by the user specified by `userID`.
      operationId: getUserPhotos
      tags:
        - users
      responses:
        200:
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  lodgings:
                    type: array
                    description: The photos posted by the specified user.
                    items:
                      $ref: '#/components/schemas/Photo'
components:
  schemas:
    Business:
      description: >
        An object representing information about a single business.
      type: object
      properties:
        ownerID:
          type: integer
          description: >
            The ID of the user who owns this business.  This is required when creating or updating a business.
          example: 16
        name:
          type: string
          description: Name of the business.
          example: American Dream Pizza
        address:
          type: string
          description: Street address of the business.
          example: 2525 NW Monroe Ave.
        city:
          type: string
          description: The city in which the business resides.
          example: Corvallis
        state:
          type: string
          description: The state in which the business resides.
          example: OR
        zip:
          type: string
          description: The zip code of the city in which the business resides.
          example: '97330'
        phone:
          type: string
          description: The business' phone number.
          example: '541-757-1713'
        category:
          type: string
          description: The business' category.
          example: 'Restaurant'
        subcategory:
          type: string
          description: The business' subcategory.
          example: 'Pizza'
        website:
          type: string
          description: The business' website URL.
          example: 'http://adpizza.com'
        email:
          type: string
          description: The business' email address.
          example: 'brian@adpizza.com'
        reviews:
          type: array
          description: >
            An array of reviews of the business.  These are only returned when fetching detailed information about a business.
          items:
            $ref: '#/components/schemas/Review'
        photos:
          type: array
          description: >
            An array of photos of the business.  These are only returned when fetching detailed information about a business.
          items:
            $ref: '#/components/schemas/Photo'
      required:
        - ownerID
        - name
        - address
        - city
        - state
        - zip
        - phone
        - category
        - subcategory
    Photo:
      description: >
        An object representing information about a photo of a business.
      type: object
      properties:
        userID:
          type: integer
          description: >
            The ID of the user associated with this photo.  This is required when creating or updating a photo.
          example: 7
        businessID:
          type: integer
          description: >
            The ID of the business associated with this photo.  This is required when creating or updating a photo.
          example: 3
        caption:
          type: string
          description: Caption text for the photo
          example: "Kale!"
        data:
          type: string
          format: byte
          description: Raw image data for the photo.
      required:
        - userID
        - businessID
        - data
    Review:
      description: >
        An object representing information about a review of a business.
      type: object
      properties:
        userID:
          type: integer
          description: >
            The ID of the user associated with this review.  This is required when creating or updating a review.
          example: 7
        businessID:
          type: integer
          description: >
            The ID of the business associated with this review.  This is required when creating or updating a review.
          example: 3
        dollars:
          type: integer
          description: A rating of the expensiveness of a business.  Must be an integer between 1 and 4.
          example: 3
          minimum: 1
          maximum: 4
        stars:
          type: number
          description: A rating of the quality of a business.  Must be an number between 0 and 5.
          example: 5
          minimum: 0
          maximum: 5
        review:
          type: string
          description: Textual business review.
          example: "Great produce!"
      required:
        - userID
        - businessID
        - dollars
        - stars
...